fix: 공통적인 의존성 부분 분리
